Move HitCountEntity and HitCountWrapper into the example that they are used in.
