--- PersistenceComparison/Assets/Scripts/HitCountEntity.cs
@@ -1,15 +1,0 @@
-﻿using Realms;
-
-public class HitCountEntity : RealmObject
-{
-    [PrimaryKey]
-    public int PrimaryKey { get; set; }
-    public int Value { get; set; }
-
-    private HitCountEntity() { }
-
-    public HitCountEntity(int primaryKey)
-    {
-        PrimaryKey = primaryKey;
-    }
-}
