--- PersistenceComparison/Assets/Scripts/HitCountWrapper.cs
@@ -1,5 +1,0 @@
-﻿[System.Serializable]
-public class HitCountWrapper
-{
-    public int value;
-}
