+++ PersistenceComparison/Assets/Scripts/RealmExample.cs
@@ -1,6 +1,20 @@
 using UnityEngine;
 using Realms;
 
+public class HitCountEntity : RealmObject
+{
+    [PrimaryKey]
+    public int PrimaryKey { get; set; }
+    public int Value { get; set; }
+
+    private HitCountEntity() { }
+
+    public HitCountEntity(int primaryKey)
+    {
+        PrimaryKey = primaryKey;
+    }
+}
+
 public class RealmExample : MonoBehaviour
 {
     // Resources:
